Lit and Unlit modes implementation (#10408)
* add isEditor in Editor2Page

* fix drop shadows

* do not show shadows in lit and unlit modes

* hide lights in unlit mode except for ambient

* temp ambient light

* move from reactor to execute loop

* fix conflict

* fix conflict

* make viewport lighting into a seperate system

* add License

* seperate primitive geom check

* use enums

* change root to origin entity

* remove primitve geom check from model check

* Drop shadow fixes

* revert

* CSM shadow and drop shadow systems need different insert orders

---------
