+++ packages/spatial/src/renderer/components/lights/AmbientLightComponent.ts
@@ -26,12 +26,13 @@ Ethereal Engine. All Rights Reserved.
 import { useEffect } from 'react'
 import { AmbientLight, Color } from 'three'
 
-import { defineComponent, useComponent } from '@etherealengine/ecs/src/ComponentFunctions'
+import { defineComponent, setComponent, useComponent } from '@etherealengine/ecs/src/ComponentFunctions'
 import { useEntityContext } from '@etherealengine/ecs/src/EntityFunctions'
 import { matches } from '@etherealengine/hyperflux'
 
-import { useDisposable } from '../../resources/resourceHooks'
-import { addObjectToGroup, removeObjectFromGroup } from './GroupComponent'
+import { useDisposable } from '../../../resources/resourceHooks'
+import { addObjectToGroup, removeObjectFromGroup } from '../GroupComponent'
+import { LightTagComponent } from './LightTagComponent'
 
 export const AmbientLightComponent = defineComponent({
   name: 'AmbientLightComponent',
@@ -65,6 +66,7 @@ export const AmbientLightComponent = defineComponent({
     const [light] = useDisposable(AmbientLight, entity)
 
     useEffect(() => {
+      setComponent(entity, LightTagComponent)
       addObjectToGroup(entity, light)
       return () => {
         removeObjectFromGroup(entity, light)
