+++ packages/spatial/src/renderer/components/lights/DirectionalLightComponent.ts
@@ -24,7 +24,7 @@ Ethereal Engine. All Rights Reserved.
 */
 
 import { useEffect } from 'react'
-import { Color, SpotLight } from 'three'
+import { BufferGeometry, Color, DirectionalLight, Float32BufferAttribute } from 'three'
 
 import {
   defineComponent,
@@ -36,35 +36,83 @@ import {
 import { useEntityContext } from '@etherealengine/ecs/src/EntityFunctions'
 import { matches, useMutableState } from '@etherealengine/hyperflux'
 
-import { LightHelperComponent } from '../../common/debug/LightHelperComponent'
-import { useDisposable } from '../../resources/resourceHooks'
-import { isMobileXRHeadset } from '../../xr/XRState'
-import { useUpdateLight } from '../functions/useUpdateLight'
-import { RendererState } from '../RendererState'
-import { addObjectToGroup, removeObjectFromGroup } from './GroupComponent'
+import { mergeBufferGeometries } from '../../../common/classes/BufferGeometryUtils'
+import { useDisposable } from '../../../resources/resourceHooks'
+import { useUpdateLight } from '../../functions/useUpdateLight'
+import { RendererState } from '../../RendererState'
+import { addObjectToGroup, removeObjectFromGroup } from '../GroupComponent'
+import { LineSegmentComponent } from '../LineSegmentComponent'
+import { LightTagComponent } from './LightTagComponent'
 
-// const ringGeom = new TorusGeometry(0.1, 0.025, 8, 12)
-// const coneGeom = new ConeGeometry(0.25, 0.5, 8, 1, true)
-// coneGeom.translate(0, -0.25, 0)
-// coneGeom.rotateX(-Math.PI / 2)
-// const geom = mergeBufferGeometries([ringGeom, coneGeom])!
-// const helperMaterial = new MeshBasicMaterial({ fog: false, transparent: true, opacity: 0.5, side: DoubleSide })
+const size = 1
+const lightPlaneGeometry = new BufferGeometry()
+lightPlaneGeometry.setAttribute(
+  'position',
+  new Float32BufferAttribute(
+    [
+      -size,
+      size,
+      0,
+      size,
+      size,
+      0,
+      size,
+      size,
+      0,
+      size,
+      -size,
+      0,
+      size,
+      -size,
+      0,
+      -size,
+      -size,
+      0,
+      -size,
+      -size,
+      0,
+      -size,
+      size,
+      0,
+      -size,
+      size,
+      0,
+      size,
+      -size,
+      0,
+      size,
+      size,
+      0,
+      -size,
+      -size,
+      0
+    ],
+    3
+  )
+)
 
-export const SpotLightComponent = defineComponent({
-  name: 'SpotLightComponent',
-  jsonID: 'EE_spot_light',
+const targetLineGeometry = new BufferGeometry()
+const t = size * 0.1
+targetLineGeometry.setAttribute(
+  'position',
+  new Float32BufferAttribute([-t, t, 0, 0, 0, 1, t, t, 0, 0, 0, 1, t, -t, 0, 0, 0, 1, -t, -t, 0, 0, 0, 1], 3)
+)
+
+const mergedGeometry = mergeBufferGeometries([targetLineGeometry, lightPlaneGeometry])
+
+export const DirectionalLightComponent = defineComponent({
+  name: 'DirectionalLightComponent',
+  jsonID: 'EE_directional_light',
 
   onInit: (entity) => {
     return {
+      light: null! as DirectionalLight,
       color: new Color(),
-      intensity: 10,
-      range: 0,
-      decay: 2,
-      angle: Math.PI / 3,
-      penumbra: 1,
+      intensity: 1,
       castShadow: false,
-      shadowBias: 0.00001,
-      shadowRadius: 1
+      shadowBias: -0.00001,
+      shadowRadius: 1,
+      cameraFar: 200
     }
   },
 
@@ -73,10 +121,7 @@ export const SpotLightComponent = defineComponent({
     if (matches.object.test(json.color) && json.color.isColor) component.color.set(json.color)
     if (matches.string.test(json.color) || matches.number.test(json.color)) component.color.value.set(json.color)
     if (matches.number.test(json.intensity)) component.intensity.set(json.intensity)
-    if (matches.number.test(json.range)) component.range.set(json.range)
-    if (matches.number.test(json.decay)) component.decay.set(json.decay)
-    if (matches.number.test(json.angle)) component.angle.set(json.angle)
-    if (matches.number.test(json.penumbra)) component.penumbra.set(json.penumbra)
+    if (matches.number.test(json.cameraFar)) component.cameraFar.set(json.cameraFar)
     if (matches.boolean.test(json.castShadow)) component.castShadow.set(json.castShadow)
     /** backwards compat */
     if (matches.number.test(json.shadowBias)) component.shadowBias.set(json.shadowBias)
@@ -87,10 +132,7 @@ export const SpotLightComponent = defineComponent({
     return {
       color: component.color.value,
       intensity: component.intensity.value,
-      range: component.range.value,
-      decay: component.decay.value,
-      angle: component.angle.value,
-      penumbra: component.penumbra.value,
+      cameraFar: component.cameraFar.value,
       castShadow: component.castShadow.value,
       shadowBias: component.shadowBias.value,
       shadowRadius: component.shadowRadius.value
@@ -101,14 +143,13 @@ export const SpotLightComponent = defineComponent({
     const entity = useEntityContext()
     const renderState = useMutableState(RendererState)
     const debugEnabled = renderState.nodeHelperVisibility
-    const spotLightComponent = useComponent(entity, SpotLightComponent)
-    const [light] = useDisposable(SpotLight, entity)
-    const lightHelper = useOptionalComponent(entity, LightHelperComponent)
+    const directionalLightComponent = useComponent(entity, DirectionalLightComponent)
+    const [light] = useDisposable(DirectionalLight, entity)
+    const lightHelper = useOptionalComponent(entity, LineSegmentComponent)
 
     useEffect(() => {
-      if (isMobileXRHeadset) return
-      light.target.position.set(1, 0, 0)
-      light.target.name = 'light-target'
+      setComponent(entity, LightTagComponent)
+      directionalLightComponent.light.set(light)
       addObjectToGroup(entity, light)
       return () => {
         removeObjectFromGroup(entity, light)
@@ -116,45 +157,31 @@ export const SpotLightComponent = defineComponent({
     }, [])
 
     useEffect(() => {
-      light.color.set(spotLightComponent.color.value)
-      if (lightHelper) lightHelper.color.set(spotLightComponent.color.value)
-    }, [spotLightComponent.color, lightHelper])
+      light.color.set(directionalLightComponent.color.value)
+      if (!lightHelper) return
+      lightHelper.color.set(directionalLightComponent.color.value)
+    }, [directionalLightComponent.color])
 
     useEffect(() => {
-      light.intensity = spotLightComponent.intensity.value
-    }, [spotLightComponent.intensity])
+      light.intensity = directionalLightComponent.intensity.value
+    }, [directionalLightComponent.intensity])
 
     useEffect(() => {
-      light.distance = spotLightComponent.range.value
-    }, [spotLightComponent.range])
+      light.shadow.camera.far = directionalLightComponent.cameraFar.value
+      light.shadow.camera.updateProjectionMatrix()
+    }, [directionalLightComponent.cameraFar])
 
     useEffect(() => {
-      light.decay = spotLightComponent.decay.value
-    }, [spotLightComponent.decay])
+      light.shadow.bias = directionalLightComponent.shadowBias.value
+    }, [directionalLightComponent.shadowBias])
 
     useEffect(() => {
-      light.angle = spotLightComponent.angle.value
-    }, [spotLightComponent.angle])
-
-    useEffect(() => {
-      light.penumbra = spotLightComponent.penumbra.value
-    }, [spotLightComponent.penumbra])
-
-    useEffect(() => {
-      light.shadow.bias = spotLightComponent.shadowBias.value
-    }, [spotLightComponent.shadowBias])
-
-    useEffect(() => {
-      light.shadow.radius = spotLightComponent.shadowRadius.value
-    }, [spotLightComponent.shadowRadius])
-
-    useEffect(() => {
-      light.castShadow = spotLightComponent.castShadow.value
-    }, [spotLightComponent.castShadow])
+      light.shadow.radius = directionalLightComponent.shadowRadius.value
+    }, [directionalLightComponent.shadowRadius])
 
     useEffect(() => {
       if (light.shadow.mapSize.x !== renderState.shadowMapResolution.value) {
-        light.shadow.mapSize.set(renderState.shadowMapResolution.value, renderState.shadowMapResolution.value)
+        light.shadow.mapSize.setScalar(renderState.shadowMapResolution.value)
         light.shadow.map?.dispose()
         light.shadow.map = null as any
         light.shadow.camera.updateProjectionMatrix()
@@ -164,10 +191,16 @@ export const SpotLightComponent = defineComponent({
 
     useEffect(() => {
       if (debugEnabled.value) {
-        setComponent(entity, LightHelperComponent, { name: 'spot-light-helper', light: light })
+        setComponent(entity, LineSegmentComponent, {
+          name: 'directional-light-helper',
+          // Clone geometry because LineSegmentComponent disposes it when removed
+          geometry: mergedGeometry?.clone(),
+          color: directionalLightComponent.color.value
+        })
       }
+
       return () => {
-        removeComponent(entity, LightHelperComponent)
+        removeComponent(entity, LineSegmentComponent)
       }
     }, [debugEnabled])
