+++ packages/spatial/src/renderer/components/lights/LightTagComponent.ts
@@ -23,30 +23,6 @@ All portions of the code written by the Ethereal Engine team are Copyright © 20
 Ethereal Engine. All Rights Reserved.
 */
 
-import { TransformComponent } from '../transform/components/TransformComponent'
-import { XRAnchorComponent } from '../xr/XRComponents'
-import { DebugRendererSystem } from './DebugRendererSystem'
-import { RenderInfoSystem } from './RenderInfoSystem'
-import { ViewportLightingSystem } from './ViewportLightingSystem'
-import { WebGLRendererSystem } from './WebGLRendererSystem'
-import { AmbientLightComponent } from './components/lights/AmbientLightComponent'
-import { DirectionalLightComponent } from './components/lights/DirectionalLightComponent'
-import { HemisphereLightComponent } from './components/lights/HemisphereLightComponent'
-import { PointLightComponent } from './components/lights/PointLightComponent'
-import { SpotLightComponent } from './components/lights/SpotLightComponent'
+import { defineComponent } from '@etherealengine/ecs'
 
-/** Components */
-export {
-  AmbientLightComponent,
-  /** Systems */
-  DebugRendererSystem,
-  DirectionalLightComponent,
-  HemisphereLightComponent,
-  PointLightComponent,
-  RenderInfoSystem,
-  SpotLightComponent,
-  TransformComponent,
-  ViewportLightingSystem,
-  WebGLRendererSystem,
-  XRAnchorComponent
-}
+export const LightTagComponent = defineComponent({ name: 'LightTag' })
