+++ packages/spatial/src/xr/XRLightProbeSystem.ts
@@ -34,10 +34,10 @@ import { defineSystem } from '@etherealengine/ecs/src/SystemFunctions'
 import { defineState, getMutableState, getState, useMutableState } from '@etherealengine/hyperflux'
 
 import { Vector3_Zero } from '../common/constants/MathConstants'
-import { DirectionalLightComponent } from '../renderer/components/DirectionalLightComponent'
+import { RendererComponent } from '../renderer/WebGLRendererSystem'
 import { addObjectToGroup } from '../renderer/components/GroupComponent'
 import { setVisibleComponent } from '../renderer/components/VisibleComponent'
-import { RendererComponent } from '../renderer/WebGLRendererSystem'
+import { DirectionalLightComponent } from '../renderer/components/lights/DirectionalLightComponent'
 import { TransformComponent } from '../transform/components/TransformComponent'
 import { XRState } from './XRState'
 import { XRSystem } from './XRSystem'
