+++ packages/ui/src/components/editor/properties/render/index.tsx
@@ -35,8 +35,8 @@ import { GLTFSnapshotState } from '@etherealengine/engine/src/gltf/GLTFState'
 import { RenderSettingsComponent } from '@etherealengine/engine/src/scene/components/RenderSettingsComponent'
 import { SourceComponent } from '@etherealengine/engine/src/scene/components/SourceComponent'
 import { State, dispatchAction } from '@etherealengine/hyperflux'
+import { DirectionalLightComponent } from '@etherealengine/spatial'
 import { NameComponent } from '@etherealengine/spatial/src/common/NameComponent'
-import { DirectionalLightComponent } from '@etherealengine/spatial/src/renderer/components/DirectionalLightComponent'
 import React, { useEffect } from 'react'
 import { useTranslation } from 'react-i18next'
 import { SiRender } from 'react-icons/si'
